moved all modal stylesheets into new styles directory
