+++ src/styles/modal/snackbar.css
@@ -1,0 +1,12 @@
+/* snackbar css */
+.snackbar {
+    display: none;
+    color: #30b130;
+    margin-left: 7%;
+    margin-top: 2%;
+    z-index: 1000;
+}
+
+.snackbar.show {
+    display: block;
+}
