+++ src/styles/modal/vehicleDropdown.css
@@ -1,0 +1,74 @@
+/* ############ 
+    CUSTOM DROPDOWN STARTS HERE 
+############ */
+.vehicle--label {
+    padding-top: 20px;
+}
+
+.vehicle--search {
+    border: 1px solid white;
+    border-bottom: 1px solid lightgray;
+    width: 421px;
+    height: 35px;
+    line-height: 27px;
+}
+
+.vehicle--search:focus {
+    outline: none;
+}
+
+.vehicles--dropdown {
+    max-width: 421px;
+    overflow: hidden;
+    /* IE and Edge */
+    -ms-overflow-style: none;
+    /* Firefox */
+    scrollbar-width: none;
+    border: 2px solid gray;
+    border-radius: 4px;
+    height: 35px;
+    transition: height 0.4s ease;
+}
+
+.vehicles--dropdown::-webkit-scrollbar {
+    display: none;
+}
+
+.vehicles--dropdown.open {
+    height: 220px;
+    overflow: scroll;
+    outline: none;
+    border-color: #6a78d1;
+    box-shadow: 0 0 8px #6a78d1;
+    background-color: #fff;
+}
+
+.vehicles--select {
+    display: flex;
+    flex-direction: column;
+    padding-bottom: 3px;
+    padding-left: 10px;
+    padding-right: 10px;
+    border-bottom: 1px solid lightgray;
+}
+
+.vehicle--title {
+    font-size: 18px;
+    color: #343a40;
+    font-weight: 500;
+    padding-top: 7px;
+}
+.vehicle--title, .vehicle--price, 
+.vin {
+    pointer-events: none;
+}
+
+.vehicles--select:hover {
+    cursor: pointer;
+    background-color: #f3f7f9d9;
+}
+
+
+/* ############ 
+    CUSTOM DROPDOWN ENDS HERE 
+############ */
